refactor: remove lib/graphql directory and move each declaration in directory where it is used
