+++ projects/safe/src/lib/components/record-modal/record-modal.component.ts
@@ -25,10 +25,7 @@ import addCustomFunctions from '../../utils/custom-functions';
 import { SafeDownloadService } from '../../services/download/download.service';
 import { SafeAuthService } from '../../services/auth/auth.service';
 import { SafeConfirmModalComponent } from '../confirm-modal/confirm-modal.component';
-import {
-  EDIT_RECORD,
-  EditRecordMutationResponse,
-} from '../../graphql/mutations';
+import { EDIT_RECORD, EditRecordMutationResponse } from './graphql/mutations';
 import { SafeSnackBarService } from '../../services/snackbar/snackbar.service';
 import { SafeFormBuilderService } from '../../services/form-builder/form-builder.service';
 import { RecordHistoryModalComponent } from '../record-history-modal/record-history-modal.component';
