+++ projects/safe/src/lib/pages/profile/profile.component.ts
@@ -6,7 +6,7 @@ import { Subscription } from 'rxjs';
 import {
   EditUserProfileMutationResponse,
   EDIT_USER_PROFILE,
-} from '../../graphql/mutations';
+} from './graphql/mutations';
 import { User } from '../../models/user.model';
 import { SafeAuthService } from '../../services/auth/auth.service';
 import { SafeSnackBarService } from '../../services/snackbar/snackbar.service';
