+++ projects/safe/src/lib/services/application/graphql/subscriptions.ts
@@ -1,34 +1,5 @@
 import { gql } from 'apollo-angular';
-import { Notification } from '../models/notification.model';
-import { Application } from '../models/application.model';
-
-/** Graphql request for listening to notifications */
-export const NOTIFICATION_SUBSCRIPTION = gql`
-  subscription NotificationSubscription {
-    notification {
-      id
-      action
-      content
-      createdAt
-      channel {
-        id
-        title
-        application {
-          id
-        }
-      }
-      seenBy {
-        id
-        name
-      }
-    }
-  }
-`;
-
-/** Model for NotificationSubscriptionResponse object */
-export interface NotificationSubscriptionResponse {
-  notification: Notification;
-}
+import { Application } from '../../../models/application.model';
 
 /** Graphql request  for listening to unlocking of applications */
 export const APPLICATION_UNLOCKED_SUBSCRIPTION = gql`
@@ -40,6 +11,11 @@ export const APPLICATION_UNLOCKED_SUBSCRIPTION = gql`
     }
   }
 `;
+
+/** Model for ApplicationUnlockedSubscriptionResponse object */
+export interface ApplicationUnlockedSubscriptionResponse {
+  applicationUnlocked: Application;
+}
 
 /** Graphql request for listening to editing of applications */
 export const APPLICATION_EDITED_SUBSCRIPTION = gql`
@@ -60,10 +36,6 @@ export const APPLICATION_EDITED_SUBSCRIPTION = gql`
   }
 `;
 
-/** Model for ApplicationUnlockedSubscriptionResponse object */
-export interface ApplicationUnlockedSubscriptionResponse {
-  applicationUnlocked: Application;
-}
 /** Model for ApplicationEditeSubscriptionResponse obejct */
 export interface ApplicationEditedSubscriptionResponse {
   applicationEdited: Application;
