+++ projects/safe/src/lib/services/grid-layout/grid-layout.service.ts
@@ -7,13 +7,13 @@ import {
   DELETE_LAYOUT,
   EditLayoutMutationResponse,
   EDIT_LAYOUT,
-} from '../../graphql/mutations';
+} from './graphql/mutations';
 import {
   GetResourceByIdQueryResponse,
   GET_GRID_RESOURCE_META,
   GetFormByIdQueryResponse,
   GET_GRID_FORM_META,
-} from '../../graphql/queries';
+} from './graphql/queries';
 import { Layout, LayoutConnection } from '../../models/layout.model';
 
 /** Default page size for layouts */
