+++ projects/safe/src/lib/services/query-builder/graphql/queries.ts
@@ -1,0 +1,151 @@
+import { gql } from 'apollo-angular';
+import { Resource } from '../../../models/resource.model';
+import { Form } from '../../../models/form.model';
+
+/**
+ * Get metadata of form / resource query definition.
+ */
+export const GET_QUERY_META_DATA = gql`
+  query GetQueryMetaData($id: ID!) {
+    form(id: $id) {
+      id
+      metadata
+    }
+    resource(id: $id) {
+      id
+      metadata
+    }
+  }
+`;
+
+/** Interface of metadata query response */
+export interface GetQueryMetaDataQueryResponse {
+  form: Form;
+  resource: Resource;
+}
+
+/** Graphql request for getting query types */
+export const GET_QUERY_TYPES = gql`
+  query GetQueryTypes {
+    __schema {
+      types {
+        name
+        kind
+        fields {
+          name
+          args {
+            name
+            type {
+              name
+              kind
+              inputFields {
+                name
+                type {
+                  name
+                  kind
+                }
+              }
+            }
+          }
+          type {
+            name
+            kind
+            fields {
+              name
+              args {
+                name
+                type {
+                  name
+                  kind
+                  inputFields {
+                    name
+                    type {
+                      name
+                      kind
+                    }
+                  }
+                }
+              }
+              type {
+                name
+                kind
+                ofType {
+                  name
+                  fields {
+                    name
+                    type {
+                      name
+                      kind
+                      ofType {
+                        name
+                      }
+                    }
+                  }
+                }
+              }
+            }
+            ofType {
+              name
+              fields {
+                name
+                type {
+                  name
+                  kind
+                  ofType {
+                    name
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+      queryType {
+        name
+        kind
+        fields {
+          name
+          args {
+            name
+            type {
+              name
+              kind
+              inputFields {
+                name
+                type {
+                  name
+                  kind
+                }
+              }
+            }
+          }
+          type {
+            name
+            kind
+            ofType {
+              name
+              fields {
+                name
+                type {
+                  name
+                  kind
+                  ofType {
+                    name
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+`;
+
+// TODO: check type of __schema
+/** Model for GetQueryTypes object */
+export interface GetQueryTypes {
+  loading: boolean;
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  __schema: any;
+}
