+++ projects/safe/src/lib/utils/custom-functions.ts
@@ -1,5 +1,5 @@
 import { Apollo } from 'apollo-angular';
-import { EDIT_RECORDS } from '../graphql/mutations';
+import { EDIT_RECORDS } from './graphql/mutations';
 import { Record } from '../models/record.model';
 import { SafeAuthService } from '../services/auth/auth.service';
